Initial work on a bb.event rewrite
Known issues:
Currently it's just plain nonworking.  After the configuration is parsed, it
dies off.
The way the dispatcher is passed into the classmethods in bb.cache.Cache is
annoying, might want to make those ordinary methods instead to avoid that, and
send the events through the cache's event dispatcher.
I also dislike passing the dispatcher into bb.build, but the way bb.build is
handled today, there aren't many options.

One thing that needs to be thought through is the firing of metadata events --
particularly with regard to the handlers from the configuration metadata.  If
we aren't careful, we could end up calling those twice, once with the config
metadata, and once with the recipe metadata -- the best bet is likely to
construct the config metadata handlers, fire off the very few events that go
directly to the config metadata, then unregister it before the recipe bits
happen, or simply unregister it in fork_off_task.

Signed-off-by: Chris Larson <chris_larson@mentor.com>