runqueue: be more pythonic in our task sorting
Rather than constructing a priority map, use sorted() to sort the list of
tasks with an appropriate key function.  While we're at it, alter the
completion scheduler to sort by the total of task weights for the recipe,
rather than shifting all recipe tasks to the position of the first seen.

Signed-off-by: Chris Larson <chris_larson@mentor.com>