Rework environment handling
Rather than clearing the environment up front, which requires that bitbake be
aware of the relative importance of particular environment variables, we can
postpone the operation, giving more control to the metadata.

- Let all environment variables flow into the metadata, rather than a
  hardcoded list
- Don't re-export environment variables implicitly at all
- Do the environment clear just before we pull exported variables from the
  metadata into the environment, when spawning the task

Of course, this means that the metadata now has to export things like PATH,
HOME, etc, which isn't necessarily a bad thing, since only it knows what
exactly it's going to be executing, and therefore, what might be needed.

An alternative to this approach would be to avoid clearing the environment for
all tasks, instead only doing so for tasks which are flagged as such in the
metadata.  Ideas:

do_compile[clearenv] = "1"

- or -

do_compile[env] = ""
do_compile[env] += "CC"

This would avoid the need to explicitly export things like DISPLAY, since we
could enable clearing for do_configure and do_compile alone.

Signed-off-by: Chris Larson <chris_larson@mentor.com>