build: only send functions to the shell process
As exported variables are already in the environment of the worker process, we
know they will be in the environment of the shell process, so emitting them to
it is unnecessary.  This greatly cleans up the set -x debug output.  We still
emit the exported variable definitions into the run. script, but only as a
debugging aid, as we no longer run the run script directly.

Signed-off-by: Chris Larson <chris_larson@mentor.com>